+++ src/EntityGraphQL/Extensions/QueryableExtensions.cs
@@ -9,21 +9,22 @@ namespace EntityGraphQL.Extensions
     /// <summary>
     /// Extension methods to allow to allow you to build queries and reuse expressions/filters
     /// </summary>
-    public static class LinqExtensions
+    public static class QueryableExtensions
     {
-        public static IEnumerable<TSource> Take<TSource>(this IEnumerable<TSource> source, int? count)
-        {
-            if (!count.HasValue)
-                return source;
-
-            return Enumerable.Take(source, count.Value);
-        }
         public static IQueryable<TSource> Take<TSource>(this IQueryable<TSource> source, int? count)
         {
             if (!count.HasValue)
                 return source;
 
             return Queryable.Take(source, count.Value);
+        }
+
+        public static IQueryable<TSource> Skip<TSource>(this IQueryable<TSource> source, int? count)
+        {
+            if (!count.HasValue)
+                return source;
+
+            return Queryable.Skip(source, count.Value);
         }
 
         public static IQueryable<TSource> WhereWhen<TSource>(this IQueryable<TSource> source, Expression<Func<TSource, bool>> wherePredicate, bool applyPredicate)
@@ -34,20 +35,6 @@ namespace EntityGraphQL.Extensions
             return source;
         }
 
-        public static IEnumerable<TSource> WhereWhen<TSource>(this IEnumerable<TSource> source, Expression<Func<TSource, bool>> wherePredicate, bool applyPredicate)
-        {
-            if (applyPredicate)
-                return Queryable.Where(source.AsQueryable(), wherePredicate).AsEnumerable();
-
-            return source;
-        }
-
-        public static IEnumerable<TSource> WhereWhen<TSource>(this IEnumerable<TSource> source, EntityQueryType<TSource> filter, bool applyPredicate)
-        {
-            if (filter.HasValue && applyPredicate)
-                return Queryable.Where(source.AsQueryable(), filter.Query);
-            return source;
-        }
         public static IQueryable<TSource> WhereWhen<TSource>(this IQueryable<TSource> source, EntityQueryType<TSource> filter, bool applyPredicate)
         {
             if (filter.HasValue && applyPredicate)