+++ src/EntityGraphQL/Schema/FieldExtensions/OffsetPaging/OffsetArgs.cs
@@ -1,0 +1,8 @@
+namespace EntityGraphQL.Schema.FieldExtensions
+{
+    public class OffsetArgs
+    {
+        public int? skip { get; set; }
+        public int? take { get; set; }
+    }
+}